# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "(◍•ᴗ•◍)❤",
        "title": "Consolidated API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/fee/outward/v1/branch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "description": "Request",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/consolidated_domain_outward_model.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "responseDetail": {
                                            "$ref": "#/definitions/consolidated_domain_outward_model.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseDesc": {
                    "type": "string"
                },
                "responseDetail": {
                    "type": "object"
                }
            }
        },
...
        "consolidated_domain_outward_model.ReqDetail": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "string"
                },
                "amountFrom": {
                    "type": "number"
                },
                "amountTo": {
                    "type": "number"
                },
                "benCountry": {
                    "type": "string"
                },
                "chargeType": {
                    "type": "string"
                },
                "cifNo": {
                    "type": "string"
                },
                "effectiveDate": {
                    "type": "string"
                },
                "exchangeRate": {
                    "type": "number"
                },
                "feeChannel": {
                    "type": "string"
                },
                "feeType": {
                    "type": "string"
                },
                "fromCCY": {
                    "type": "string"
                },
                "orderingType": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "searchPayInFull": {
                    "type": "string"
                },
                "toCCY": {
                    "type": "string"
                },
                "transactionType": {
                    "type": "string"
                }
            }
        },
        "consolidated_domain_outward_model.Request": {
            "type": "object",
            "properties": {
                "channelID": {
                    "type": "string"
                },
                "refId": {
                    "type": "string"
                },
                "reqDetail": {
                    "$ref": "#/definitions/consolidated_domain_outward_model.ReqDetail"
                },
                "transDate": {
                    "type": "string"
                },
                "transTime": {
                    "type": "string"
                }
            }
        },
        "consolidated_domain_outward_model.Response": {
            "type": "object",
            "properties": {
                "feeAmount": {
                    "type": "number"
                },
                "feeCCY": {
                    "type": "string"
                },
                "feeCategory": {
                    "type": "string"
                },
                "feeRefID": {
                    "type": "string"
                },
                "feeType": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Bearer token"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}